avoid code duplication in gsa_mixer/cli.py and bivar_mixer/cli.py
