+++ precimed/gsa_mixer/cli.py
@@ -17,6 +17,8 @@ from pathlib import Path
 from scipy.sparse import csr_matrix, coo_matrix
 
 from common.libbgmg import LibBgmg, LibBgmgAnnot
+from common.libbgmg import tagvec_as_snpvec
+
 from .utils import AnnotUnivariateParams
 from .utils import _params_to_dict
 from .utils import _dict_to_params
@@ -24,115 +26,12 @@ from .utils import AnnotUnivariateParametrization
 
 from common.utils import NumpyEncoder
 from common.libbgmg import _cost_calculator_sampling
-
-MASTHEAD = "***********************************************************************\n"
-MASTHEAD += "* (c) 2016-2024 MiXeR software - Univariate and Bivariate Causal Mixture for GWAS\n"
-MASTHEAD += "* Norwegian Centre for Mental Disorders Research / University of Oslo\n"
-MASTHEAD += "* Center for Multimodal Imaging and Genetics / UCSD\n"
-MASTHEAD += "* GNU General Public License v3\n"
-MASTHEAD += "***********************************************************************\n"
-
+from common.utils_cli import check_input_file, make_ranges
+from common.utils_cli import fix_and_validate_args, convert_args_to_libbgmg_options
 
 GO_EXTEND_BP_DEFAULT = 10000
 
 CHR_OFFSET = np.int64(1e10)
-
-def check_input_file(args, argname, chri=None):
-    arg_dict = vars(args)
-    if (argname in arg_dict) and arg_dict[argname]:
-        argval = arg_dict[argname]
-        if chri: argval=argval.replace('@', str(chri))
-        if not os.path.isfile(argval): raise ValueError("Input file --{a} does not exist: {f}".format(a=argname, f=argval))
-
-def fix_chr2use_arg(args, argname="chr2use"):
-    arg_dict = vars(args)
-    if argname not in arg_dict : return
-    if not arg_dict[argname]: arg_dict[argname] = []; return
-    chr2use_arg = arg_dict[argname]
-    chr2use = []
-    if chr2use_arg is not None:
-        for a in chr2use_arg.split(","):
-            if "-" in a:
-                start, end = [int(x) for x in a.split("-")]
-                chr2use += [str(x) for x in range(start, end+1)]
-            else:
-                chr2use.append(a.strip())
-        if np.any([not x.isdigit() for x in chr2use]): raise ValueError('Chromosome labels must be integer')
-    arg_dict[argname] = [int(x) for x in chr2use]
-
-def make_ranges(args_exclude_ranges):
-    # Interpret --exclude-ranges input
-    ChromosomeRange = collections.namedtuple('ChromosomeRange', ['chr', 'from_bp', 'to_bp'])
-    exclude_ranges = []
-    if args_exclude_ranges is not None:
-        for exclude_range in args_exclude_ranges:
-            try:
-                exclude_range=exclude_range.strip().upper().replace('CHR','')
-                if exclude_range == '[]': continue
-                if exclude_range == 'MHC': exclude_range='6:25-35MB'
-                if exclude_range == 'APOE': exclude_range = '19:45-46MB'
-                if exclude_range.endswith('KB'): factor=1e3
-                elif exclude_range.endswith('MB'): factor=1e6
-                else: factor=1
-                exclude_range=exclude_range.replace('KB', '').replace('MB', '')
-                chr_from_to = [int(x) for x in exclude_range.replace(':', ' ').replace('-', ' ').split()[:3]]
-                chr_from_to[1] *= factor
-                chr_from_to[2] *= factor
-                range = ChromosomeRange._make(chr_from_to)
-            except Exception as e:
-                raise(ValueError('Unable to interpret exclude range "{}", chr:from-to format is expected.'.format(exclude_range)))
-            exclude_ranges.append(range)
-    return exclude_ranges
-
-def fix_and_validate_args(args):
-    # here is what you need to know about how --seed  is handled:
-    # 1. If it optional. When not specified, we will use np.rando.randint to generate a random seed (the the line below this comment)
-    # 2. There are several places in this code where we generate random numbers using numpy library. Therefore, we use args.seed to initialize np.random.seed.
-    # 3. We pass args.seed to MiXeR using set_options("seed", args.seed), see convert_args_to_libbgmg_options() function.
-    #    This will ensure that random prunning uses consistent seed.
-    # 4. In ADAM algorithm, each cost function evaluation starts with it's own seed (to respect the stochastic nature of the algorithm)
-    #    This is done in line "libbgmg.set_option('seed', np.random.randint(np.iinfo(np.int32).max))".
-    # 5. --diffevo-fast-repeats also rely on seed to ensure different path of the differential evolution algorihtm.
-    if 'seed' in args:
-        if args.seed is None: args.seed = np.random.randint(np.iinfo(np.int32).max)
-        np.random.seed(args.seed)
-
-    fix_chr2use_arg(args, argname="chr2use")
-    if 'trait1_file' in args:
-        if '@' in args.trait1_file:
-            for chri in args.chr2use:
-                    check_input_file(args, 'trait1_file', chri)
-        else:
-            check_input_file(args, 'trait1_file')
-    check_input_file(args, 'trait2_file')
-    check_input_file(args, 'trait1_params_file')
-    check_input_file(args, 'trait2_params_file')
-    check_input_file(args, 'load-params-file')
-
-    if ('analysis' in args) and (args.analysis == ''):
-        if not args.load_params_file:
-            if not args.trait1_params_file: raise ValueError('--trait1-params-file or --load-params-file is required ')
-            if not args.trait2_params_file: raise ValueError('--trait2-params-file or --load-params-file is required ')
-
-    arg_dict = vars(args)
-    for chri in arg_dict["chr2use"]:
-        check_input_file(args, 'bim-file', chri)
-        check_input_file(args, 'ld-file', chri)
-        check_input_file(args, 'annot-file', chri)
-
-def convert_args_to_libbgmg_options(args):
-    libbgmg_options = {
-        'r2min': args.r2min if ('r2min' in args) else None,
-        'kmax': args.kmax[0] if ('kmax' in args) else None, 
-        'kmax_pdf': args.kmax_pdf[0] if ('kmax_pdf' in args) else None,
-        'threads': args.threads[0] if ('threads' in args) else None,
-        'seed': args.seed if ('seed' in args) else None,
-        'cubature_rel_error': args.cubature_rel_error if ('cubature_rel_error' in args) else None,
-        'cubature_max_evals': args.cubature_max_evals if ('cubature_max_evals' in args) else None,
-        'z1max': args.z1max if ('z1max' in args) else None,
-        'z2max': args.z2max if ('z2max' in args) else None, 
-    }
-    return [(k, v) for k, v in libbgmg_options.items() if v is not None ]
 
 def set_weights(args, libbgmg, randprune_n):
     if 'weights_file' in args:
@@ -165,18 +64,6 @@ def set_weights(args, libbgmg, randprune_n):
 
     if (randprune_n is not None) and (randprune_n != 0):
         libbgmg.set_weights_randprune(randprune_n, args.randprune_r2, 0, not args.disable_inverse_ld_score_weights, "", "")
-
-# https://stackoverflow.com/questions/27433316/how-to-get-argparse-to-read-arguments-from-a-file-with-an-option-rather-than-pre
-class LoadFromFile (argparse.Action):
-    def __call__ (self, parser, namespace, values, option_string=None):
-        with values as f:
-            contents = '\n'.join([x for x in f.readlines() if (not x.strip().startswith('#'))])
-
-        data = parser.parse_args(contents.split(), namespace=namespace)
-        for k, v in vars(data).items():
-            if v and k != option_string.lstrip('-'):
-                setattr(namespace, k, v)
-
 
 def parser_add_argument_bim_file(parser):
     parser.add_argument("--bim-file", type=str, default=None, help="plink bim file (required argument); "
@@ -305,45 +192,6 @@ def parser_ld_add_arguments(func, parser):
     parser.add_argument('--ld-window-kb', type=float, default=0, help="limit window similar to --ld-window-kb in 'plink r2'; 0 will disable this constraint (default: %(default)s); either ld-window-kb or ld-window argument must be provided")
     parser.add_argument('--ld-window', type=int, default=0, help="limit window similar to --ld-window in 'plink r2'; 0 will disable this constraint (default: %(default)s); either ld-window-kb or ld-window argument must be provided")
     parser.set_defaults(func=func)
-
-def initialize_libbgmg_logging(args):
-    libbgmg = LibBgmg(args.lib)
-    log_fname = args.log if args.log else args.out + '.log'
-    if os.path.exists(log_fname): os.system(f'rm {log_fname}')
-    libbgmg.init_log(log_fname)
-    log_header(args, sys.argv[1], libbgmg)
-    libbgmg.dispose()
-    return libbgmg
-
-def generate_args_parser(__version__=None):
-    parser = argparse.ArgumentParser(description=f"MiXeR v{__version__}: Univariate and Bivariate Causal Mixture for GWAS.")
-    parser.add_argument('--version', action='version', version=f'MiXeR v{__version__}')
-
-    parent_parser = argparse.ArgumentParser(add_help=False)
-    parent_parser.add_argument('--argsfile', type=open, action=LoadFromFile, default=None, help=argparse.SUPPRESS) #"file with additional command-line arguments, e.g. those that are across all of your mixer.py runs (--lib, --bim-file and --ld-file)")
-    parent_parser.add_argument("--out", type=str, default="mixer", help="prefix for the output files, such as <out>.json (default: %(default)s);")
-    parent_parser.add_argument("--lib", type=str, default="libbgmg.so", help="path to libbgmg.so plugin (default: %(default)s); can be also specified via BGMG_SHARED_LIBRARY env variable.")
-    parent_parser.add_argument("--log", type=str, default=None, help="file to output log (default: <out>.log); NB! if --log points to an existing file the new lines will be appended to it at the end of the file.")
-
-    subparsers = parser.add_subparsers()
-    parser_ld_add_arguments(func=execute_ld_parser, parser=subparsers.add_parser("ld", parents=[parent_parser], help='prepare files with linkage disequilibrium information'))
-    parser_snps_add_arguments(func=execute_snps_parser, parser=subparsers.add_parser("snps", parents=[parent_parser], help='generate random sets of SNPs'))
-    parser_plsa_add_arguments(func=execute_plsa_parser, parser=subparsers.add_parser("plsa", parents=[parent_parser], help='fit PLSA model'))
-    parser_split_sumstats_add_arguments(func=execute_split_sumstats_parser, parser=subparsers.add_parser("split_sumstats", parents=[parent_parser], help='split summary statistics file per-chromosome'))
-
-    return parser
-
-def log_header(args, subparser_name, lib):
-    defaults = vars(generate_args_parser().parse_args([subparser_name]))
-    opts = vars(args)
-    non_defaults = [x for x in opts.keys() if opts[x] != defaults[x]]
-    header = MASTHEAD
-    header += "Call: \n"
-    header += './mixer.py {} \\\n'.format(subparser_name)
-    options = ['\t--'+x.replace('_','-')+' '+(' '.join([str(y) for y in opts[x]]) if isinstance(opts[x], list) else str(opts[x])).replace('\t', '\\t')+' \\' for x in non_defaults]
-    header += '\n'.join(options).replace('True','').replace('False','')
-    header = header[0:-1]+'\n'
-    lib.log_message(header)
 
 def load_params_file(fname_or_params, libbgmg, args):
     if isinstance(fname_or_params, AnnotUnivariateParams):
@@ -506,15 +354,7 @@ def apply_plsa_univariate_fit_sequence(args, params, results, libbgmg_vec, trait
 
     return params
 
-# helper function to debug non-json searizable types...
-def print_types(results, libbgmg):
-    if isinstance(results, dict):
-        for k, v in results.items():
-            libbgmg.log_message('{}: {}'.format(k, type(v)))
-            print_types(v, libbgmg)
-
 def execute_ld_parser(args):
-    initialize_libbgmg_logging(args)
     libbgmg = LibBgmg(args.lib)
     libbgmg.calc_ld_matrix(args.bfile, args.out, args.r2min, args.ldscore_r2min, args.ld_window, args.ld_window_kb)
     libbgmg.log_message('Done')
@@ -645,7 +485,6 @@ def initialize_mixer_plugin(args, context_id, trait1_file, trait2_file, randprun
     return libbgmg
 
 def execute_snps_parser(args):
-    initialize_libbgmg_logging(args)
     fix_and_validate_args(args)
 
     libbgmg = initialize_mixer_plugin(args, 0, "", "", None, None, None, args.chr2use, '@')
@@ -697,12 +536,6 @@ def init_results_struct(args, libbgmg_vec=None):
         results['options']['time_started'] = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
         results['ci'] = {}
     return results
-
-def tagvec_as_snpvec(tagvec, defvec):
-        snpvec = np.empty(defvec.shape, dtype=float)
-        snpvec[:] = np.nan
-        snpvec[defvec] = tagvec
-        return snpvec
 
 def find_per_snp_information_univariate(libbgmg_vec, params):
     # Save a file with the following information
@@ -827,7 +660,7 @@ def compute_geneset_enrichment(go_file_test, args, exclude_ranges, libbgmg_vec, 
     return df_enrich
 
 def execute_plsa_parser(args):
-    libbgmg = initialize_libbgmg_logging(args)
+    libbgmg = LibBgmg(args.lib)
     fix_and_validate_args(args)
     if (args.load_params_file is None) and (args.adam_disable):
         raise ValueError('--load-params-file is required for --adam-disable')
